Invert mod_websocket_method_handler() conditionals
The main handler function used a lot of nested if blocks, which meant
most of the code was indented several times. Combined with an
80-character column limit, it was getting very difficult to read. This
patch inverts many of the conditionals to unindent the code.

This introduced a vestigal scope block before the declaration of the
WebSocketState and WebSocketServer. It probably indicates that the block
should be pulled out into its own helper, but more work needs to be done
before that makes sense.

This patch is intended to be refactor-only; there should be no changes
to the logic.