Generate a Makefile from configure
Makefile.in contains the Makefile template, which is filled in by
configure when it is run. For now, I'm purposefully not using Automake,
since I'm using apxs for the actual building and installing.

The Makefile also contains some helper rules (start/stop/restart for
quick access to apachectl, and check for testing).