Search for and use httpd directly in tests
Some distributions (e.g. Ubuntu) put so much hardcoded logic into
apachectl that we cannot use it to launch a standalone server reliably.
Switch to using httpd directly (which requires an additional bit of
configure logic) for tests. We may have to find a replacement for the
standard envvars fixup that apachectl did for us.