Add CMake support for Unix systems
This is a squashed commit based on toniotelli's pull request #26:

    https://github.com/jchampio/apache-websocket/pull/26

It adds support for CMake on *nix systems. Previously, only Windows was
supported.

Note that the FindAPR/FindAPACHE CMake modules that were part of the
original change have been removed, since we're not really sure who the
authors are. They will be reimplemented in a followup commit.