Add find modules and use apxs for CMake installs
By using apxs for installation and module activation, we don't need to
keep track of system-dependent conventions ourselves.

The APR find module should be able to deal with both APR[Util] 1.x and
APR 2 systems. The Apache find module is more bare-bones, and simply
finds the include directory and the apxs tool for installation.